# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "h13 API - V1",
    "description": "You should migrate your app to NestJS. \n\n<b>For this migration step you SHOULDN'T migrate this functionality:</b>\n<ul>\n    <li>\n    entire auth flow (including devices), basic auth and bearer auth\n    </li>\n    <li>\n    validation\n    </li>\n    <li>\n    create/update/delete for comments (because for this operations we need bearer auth), but read requests should work for comments that already are in DB\n    </li>\n    <li>\n    create/update/delete for likes for comments and for posts \n    (because for this operations we need bearer auth).\n    But read requests should work for likes that already are in DB when we read posts and comments\n    </li>\n</ul>\n\n<br />\n\nYour should implement all other functionality using NestJS. New Tests will help you to check yourself.",
    "version": "h13"
  },
  "paths": {
...
    "/hometask_13/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Returns all users",
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "searchLoginTerm",
            "in": "query",
            "description": "Search term for user Login: Login should contains this term in any position",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "searchEmailTerm",
            "in": "query",
            "description": "Search term for user Email: Email should contains this term in any position",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h05.db.UserViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h05.db.UserViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h05.db.UserViewModel>"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Add new user to the system",
        "requestBody": {
          "description": "Data for constructing new user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h05.db.UserInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h05.db.UserInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h05.db.UserInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created user",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h05.db.UserViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h05.db.UserViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h05.db.UserViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/hometask_13/api/users/{id}": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user specified by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "If specified user is not exists"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Paginator<h05.db.UserViewModel>": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "pagesCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h05.db.UserViewModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "h04.SortDirections": {
        "enum": ["asc", "desc"],
        "type": "string"
      },
      "h05.db.UserInputModel": {
        "required": ["email", "login", "password"],
        "type": "object",
        "properties": {
          "login": {
            "maxLength": 10,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_-]*$",
            "type": "string",
            "description": "must be unique"
          },
          "password": {
            "maxLength": 20,
            "minLength": 6,
            "type": "string"
          },
          "email": {
            "pattern": "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$",
            "type": "string",
            "description": "must be unique",
            "example": "example@example.dev"
          }
        },
        "additionalProperties": false
      },
      "h05.db.UserViewModel": {
        "required": ["createdAt", "email", "id", "login"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token **_only_**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "refreshToken": {
        "type": "apiKey",
        "description": "JWT refreshToken inside cookie. Must be correct, and must not expire",
        "name": "refreshToken",
        "in": "cookie"
      }
    }
  }
}